test: MemberCustomerServiceTest 초안 작성 및 x,y 좌표 Double로 변경 #135

+++ src/main/java/poomasi/domain/member/_profile/entity/MemberProfile.java
@@ -48,10 +48,10 @@ public class MemberProfile {
     private String addressDetail;
 
     @Column(nullable = true, length = 255)
-    private Long coordinateX;
+    private Double coordinateX;
 
     @Column(nullable = true, length = 255)
-    private Long coordinateY;
+    private Double coordinateY;
 
     @PrePersist
     public void prePersist() {
@@ -72,8 +72,8 @@ public class MemberProfile {
     public void setAddress(
             String defaultAddress,
             String addressDetail,
-            Long coordinateX,
-            Long coordinateY) {
+            Double coordinateX,
+            Double coordinateY) {
         if (defaultAddress != null) this.defaultAddress = defaultAddress;
         if (addressDetail != null) this.addressDetail = addressDetail;
         if (coordinateX != null) this.coordinateX = coordinateX;
